+++ interface_config.js
@@ -153,19 +153,7 @@ var interfaceConfig = {
      *
      * @type {boolean}
      */
-    VIDEO_QUALITY_LABEL_DISABLED: false,
-
-    /**
-     * This is a temporary feature flag used to gate access to the toolbox so it
-     * can be developed through smaller changesets and set to false if bad bugs
-     * are found. This feature flag will be removed at some point, as well as
-     * the old toolbox. This new toolbox will be horizontal and the previous
-     * feature of supporting menu button ordering through interfaceConfig will
-     * be removed. Support for configuring which buttons display will remain.
-     *
-     * @type {boolean}
-     */
-    _USE_NEW_TOOLBOX: true
+    VIDEO_QUALITY_LABEL_DISABLED: false
 
     /**
      * Specify custom URL for downloading android mobile app.
