+++ react/features/toolbox/components/index.js
@@ -2,6 +2,5 @@ export { default as ToolbarButton } from './ToolbarButton';
 export { default as ToolbarButtonV2 } from './ToolbarButtonV2';
 export { default as ToolbarButtonWithDialog }
     from './ToolbarButtonWithDialog';
-export { default as Toolbox } from './Toolbox';
 export { default as ToolboxFilmstrip } from './ToolboxFilmstrip';
-export { default as ToolboxV2 } from './ToolboxV2';
+export { default as Toolbox } from './Toolbox';
