--- react/features/toolbox/defaultToolbarButtons.native.js
@@ -1,1 +1,0 @@
-export default undefined;
