+++ react/features/toolbox/reducer.js
@@ -8,7 +8,6 @@ import {
     SET_DEFAULT_TOOLBOX_BUTTONS,
     SET_SUBJECT,
     SET_SUBJECT_SLIDE_IN,
-    SET_TOOLBAR_BUTTON,
     SET_TOOLBAR_HOVERED,
     SET_TOOLBOX_ALWAYS_VISIBLE,
     SET_TOOLBOX_ENABLED,
@@ -16,7 +15,6 @@ import {
     SET_TOOLBOX_TIMEOUT_MS,
     SET_TOOLBOX_VISIBLE
 } from './actionTypes';
-import getDefaultButtons from './defaultToolbarButtons';
 
 declare var interfaceConfig: Object;
 
@@ -161,9 +159,6 @@ ReducerRegistry.register(
                 subjectSlideIn: action.subjectSlideIn
             };
 
-        case SET_TOOLBAR_BUTTON:
-            return _setToolbarButton(state, action);
-
         case SET_TOOLBAR_HOVERED:
             return {
                 ...state,
@@ -204,46 +199,3 @@ ReducerRegistry.register(
 
         return state;
     });
-
-/**
- * Reduces the redux action {@code SET_TOOLBAR_BUTTON} in the feature toolbox.
- *
- * @param {Object} state - The redux state.
- * @param {Object} action - The redux action of type {@code SET_TOOLBAR_BUTTON}.
- * @param {Object} action.button - Object describing toolbar button.
- * @param {Object} action.buttonName - The name of the button.
- * @private
- * @returns {Object}
- */
-function _setToolbarButton(state, { button, buttonName }): Object {
-    // XXX getDefaultButtons, defaultToolbarButtons, SET_TOOLBAR_BUTTON are
-    // abstractions fully implemented on Web only.
-    const buttons = getDefaultButtons && getDefaultButtons();
-    const buttonDefinition = buttons && buttons[buttonName];
-
-    // We don't need to update if the button shouldn't be displayed
-    if (!buttonDefinition || !buttonDefinition.isDisplayed()) {
-        return state;
-    }
-
-    const { primaryToolbarButtons, secondaryToolbarButtons } = state;
-    let selectedButton = primaryToolbarButtons.get(buttonName);
-    let place = 'primaryToolbarButtons';
-
-    if (!selectedButton) {
-        selectedButton = secondaryToolbarButtons.get(buttonName);
-        place = 'secondaryToolbarButtons';
-    }
-
-    selectedButton = {
-        ...selectedButton,
-        ...button
-    };
-
-    const updatedToolbar = state[place].set(buttonName, selectedButton);
-
-    return {
-        ...state,
-        [place]: new Map(updatedToolbar)
-    };
-}
